requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
redis[hiredis]==5.0.1
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
slowapi==0.1.9
openai==1.51.0
bcrypt==4.1.3
pytest==8.0.0
pytest-asyncio==0.23.5
aiosqlite==0.20.0
